backend/requirements.txt
# ── Core API ─────────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20
aiofiles==24.1.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.1.0
structlog==25.1.0
rich==13.9.4
faker==37.1.0
google-genai==0.6.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1

# ── OPTIONAL: Heavy ML stack (not imported by any current app code) ───────────
# Install only if you need real ML inference (not needed with synthetic data):
scikit-learn>=1.4.1
# numpy>=1.26.4
pandas>=2.2.1
# scipy>=1.13.0
# xgboost>=2.0.3
# shap>=0.44.1
# lifetimes==0.11.3
# transformers>=4.38.2
# celery>=5.3.6
# redis>=5.0.3
# optuna>=3.5.0
# mlflow>=2.11.3
# sqlalchemy>=2.0.29
# alembic>=1.13.1
# aiosqlite>=0.20.0
